refactor: #129 - Comment 컴포넌트 common 폴더로 이동
