+++ src/components/index.ts
@@ -15,6 +15,6 @@ export { default as DropdownItem } from './common/Dropdown/DropdownItem'
 export { default as Dropdown } from './common/Dropdown/Dropdown'
 export { default as SpaceMemberList } from './common/SpaceMemberList/SpaceMemberList'
 export { default as Sidebar } from './common/Sidebar/Sidebar'
-export { default as Comment } from './Comment/Comment'
+export { default as Comment } from './common/Comment/Comment'
 export { default as Notification } from './common/Notification/Notification'
 export { default as SpaceList } from './SpaceList/SpaceList'
